+++ opentelemetry-instrumentation/src/opentelemetry/instrumentation/utils.py
@@ -12,9 +12,24 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from typing import Dict, Sequence
+
 from wrapt import ObjectProxy
 
 from opentelemetry.trace.status import StatusCanonicalCode
+
+
+def extract_attributes_from_object(
+    obj: any, attributes: Sequence[str], existing: Dict[str, str] = None
+) -> Dict[str, str]:
+    extracted = {}
+    if existing:
+        extracted.update(existing)
+    for attr in attributes:
+        value = getattr(obj, attr, None)
+        if value is not None:
+            extracted[attr] = str(value)
+    return extracted
 
 
 def http_status_to_canonical_code(