+++ opentelemetry-sdk/src/opentelemetry/sdk/trace/__init__.py
@@ -18,12 +18,11 @@ import atexit
 import concurrent.futures
 import json
 import logging
-import random
 import threading
 import traceback
 from collections import OrderedDict
 from contextlib import contextmanager
-from types import TracebackType
+from types import MappingProxyType, TracebackType
 from typing import (
     Any,
     Callable,
@@ -45,14 +44,18 @@ from opentelemetry.sdk.util import BoundedDict, BoundedList
 from opentelemetry.sdk.util.instrumentation import InstrumentationInfo
 from opentelemetry.trace import SpanContext
 from opentelemetry.trace.propagation import SPAN_KEY
-from opentelemetry.trace.status import Status, StatusCanonicalCode
+from opentelemetry.trace.status import (
+    EXCEPTION_STATUS_FIELD,
+    Status,
+    StatusCanonicalCode,
+)
 from opentelemetry.util import time_ns, types
 
 logger = logging.getLogger(__name__)
 
-MAX_NUM_ATTRIBUTES = 32
-MAX_NUM_EVENTS = 128
-MAX_NUM_LINKS = 32
+MAX_NUM_ATTRIBUTES = 1000
+MAX_NUM_EVENTS = 1000
+MAX_NUM_LINKS = 1000
 VALID_ATTR_VALUE_TYPES = (bool, str, int, float)
 
 
@@ -65,7 +68,11 @@ class SpanProcessor:
     in the same order as they were registered.
     """
 
-    def on_start(self, span: "Span") -> None:
+    def on_start(
+        self,
+        span: "Span",
+        parent_context: Optional[context_api.Context] = None,
+    ) -> None:
         """Called when a :class:`opentelemetry.trace.Span` is started.
 
         This method is called synchronously on the thread that starts the
@@ -73,6 +80,7 @@ class SpanProcessor:
 
         Args:
             span: The :class:`opentelemetry.trace.Span` that just started.
+            parent_context: The parent context of the span that just started.
         """
 
     def on_end(self, span: "Span") -> None:
@@ -121,9 +129,13 @@ class SynchronousMultiSpanProcessor(SpanProcessor):
         with self._lock:
             self._span_processors = self._span_processors + (span_processor,)
 
-    def on_start(self, span: "Span") -> None:
+    def on_start(
+        self,
+        span: "Span",
+        parent_context: Optional[context_api.Context] = None,
+    ) -> None:
         for sp in self._span_processors:
-            sp.on_start(span)
+            sp.on_start(span, parent_context=parent_context)
 
     def on_end(self, span: "Span") -> None:
         for sp in self._span_processors:
@@ -189,17 +201,26 @@ class ConcurrentMultiSpanProcessor(SpanProcessor):
             self._span_processors = self._span_processors + (span_processor,)
 
     def _submit_and_await(
-        self, func: Callable[[SpanProcessor], Callable[..., None]], *args: Any
+        self,
+        func: Callable[[SpanProcessor], Callable[..., None]],
+        *args: Any,
+        **kwargs: Any
     ):
         futures = []
         for sp in self._span_processors:
-            future = self._executor.submit(func(sp), *args)
+            future = self._executor.submit(func(sp), *args, **kwargs)
             futures.append(future)
         for future in futures:
             future.result()
 
-    def on_start(self, span: "Span") -> None:
-        self._submit_and_await(lambda sp: sp.on_start, span)
+    def on_start(
+        self,
+        span: "Span",
+        parent_context: Optional[context_api.Context] = None,
+    ) -> None:
+        self._submit_and_await(
+            lambda sp: sp.on_start, span, parent_context=parent_context
+        )
 
     def on_end(self, span: "Span") -> None:
         self._submit_and_await(lambda sp: sp.on_end, span)
@@ -336,6 +357,10 @@ def _filter_attribute_values(attributes: types.Attributes):
                 attributes.pop(attr_key)
 
 
+def _create_immutable_attributes(attributes):
+    return MappingProxyType(attributes.copy() if attributes else {})
+
+
 class Span(trace_api.Span):
     """See `opentelemetry.trace.Span`.
 
@@ -346,7 +371,7 @@ class Span(trace_api.Span):
         name: The name of the operation this span represents
         context: The immutable span context
         parent: This span's parent's `opentelemetry.trace.SpanContext`, or
-            null if this is a root span
+            None if this is a root span
         sampler: The sampler used to create this span
         trace_config: TODO
         resource: Entity producing telemetry
@@ -356,6 +381,11 @@ class Span(trace_api.Span):
         span_processor: `SpanProcessor` to invoke when starting and ending
             this `Span`.
     """
+
+    def __new__(cls, *args, **kwargs):
+        if cls is Span:
+            raise TypeError("Span must be instantiated via a tracer.")
+        return super().__new__(cls)
 
     def __init__(
         self,
@@ -399,6 +429,10 @@ class Span(trace_api.Span):
         if events:
             for event in events:
                 _filter_attribute_values(event.attributes)
+                # pylint: disable=protected-access
+                event._attributes = _create_immutable_attributes(
+                    event.attributes
+                )
                 self.events.append(event)
 
         if links is None:
@@ -447,6 +481,8 @@ class Span(trace_api.Span):
     def _format_attributes(attributes):
         if isinstance(attributes, BoundedDict):
             return attributes._dict  # pylint: disable=protected-access
+        if isinstance(attributes, MappingProxyType):
+            return attributes.copy()
         return attributes
 
     @staticmethod
@@ -510,7 +546,7 @@ class Span(trace_api.Span):
 
         return json.dumps(f_span, indent=indent)
 
-    def get_context(self):
+    def get_span_context(self):
         return self.context
 
     def set_attribute(self, key: str, value: types.AttributeValue) -> None:
@@ -557,8 +593,7 @@ class Span(trace_api.Span):
         timestamp: Optional[int] = None,
     ) -> None:
         _filter_attribute_values(attributes)
-        if not attributes:
-            attributes = self._new_attributes()
+        attributes = _create_immutable_attributes(attributes)
         self._add_event(
             Event(
                 name=name,
@@ -567,7 +602,11 @@ class Span(trace_api.Span):
             )
         )
 
-    def start(self, start_time: Optional[int] = None) -> None:
+    def start(
+        self,
+        start_time: Optional[int] = None,
+        parent_context: Optional[context_api.Context] = None,
+    ) -> None:
         with self._lock:
             if not self.is_recording():
                 return
@@ -579,7 +618,7 @@ class Span(trace_api.Span):
         if has_started:
             logger.warning("Calling start() on a started span.")
             return
-        self.span_processor.on_start(self)
+        self.span_processor.on_start(self, parent_context=parent_context)
 
     def end(self, end_time: Optional[int] = None) -> None:
         with self._lock:
@@ -663,22 +702,11 @@ class Span(trace_api.Span):
         )
 
 
-def generate_span_id() -> int:
-    """Get a new random span ID.
+class _Span(Span):
+    """Protected implementation of `opentelemetry.trace.Span`.
 
-    Returns:
-        A random 64-bit int for use as a span ID
+    This constructor should only be used internally.
     """
-    return random.getrandbits(64)
-
-
-def generate_trace_id() -> int:
-    """Get a new random trace ID.
-
-    Returns:
-        A random 128-bit int for use as a trace ID
-    """
-    return random.getrandbits(128)
 
 
 class Tracer(trace_api.Tracer):
@@ -701,45 +729,48 @@ class Tracer(trace_api.Tracer):
     def start_as_current_span(
         self,
         name: str,
-        parent: trace_api.ParentSpan = trace_api.Tracer.CURRENT_SPAN,
+        context: Optional[context_api.Context] = None,
         kind: trace_api.SpanKind = trace_api.SpanKind.INTERNAL,
         attributes: types.Attributes = None,
         links: Sequence[trace_api.Link] = (),
+        record_exception: bool = True,
     ) -> Iterator[trace_api.Span]:
-        span = self.start_span(name, parent, kind, attributes, links)
-        return self.use_span(span, end_on_exit=True)
+        span = self.start_span(name, context, kind, attributes, links)
+        return self.use_span(
+            span, end_on_exit=True, record_exception=record_exception
+        )
 
     def start_span(  # pylint: disable=too-many-locals
         self,
         name: str,
-        parent: trace_api.ParentSpan = trace_api.Tracer.CURRENT_SPAN,
+        context: Optional[context_api.Context] = None,
         kind: trace_api.SpanKind = trace_api.SpanKind.INTERNAL,
         attributes: types.Attributes = None,
         links: Sequence[trace_api.Link] = (),
         start_time: Optional[int] = None,
         set_status_on_exception: bool = True,
     ) -> trace_api.Span:
-        if parent is Tracer.CURRENT_SPAN:
-            parent = trace_api.get_current_span()
 
-        parent_context = parent
-        if isinstance(parent_context, trace_api.Span):
-            parent_context = parent.get_context()
+        parent_span_context = trace_api.get_current_span(
+            context
+        ).get_span_context()
 
-        if parent_context is not None and not isinstance(
-            parent_context, trace_api.SpanContext
+        if parent_span_context is not None and not isinstance(
+            parent_span_context, trace_api.SpanContext
         ):
-            raise TypeError("parent must be a Span, SpanContext or None.")
+            raise TypeError(
+                "parent_span_context must be a SpanContext or None."
+            )
 
-        if parent_context is None or not parent_context.is_valid:
-            parent = parent_context = None
-            trace_id = generate_trace_id()
+        if parent_span_context is None or not parent_span_context.is_valid:
+            parent_span_context = None
+            trace_id = self.source.ids_generator.generate_trace_id()
             trace_flags = None
             trace_state = None
         else:
-            trace_id = parent_context.trace_id
-            trace_flags = parent_context.trace_flags
-            trace_state = parent_context.trace_state
+            trace_id = parent_span_context.trace_id
+            trace_flags = parent_span_context.trace_flags
+            trace_state = parent_span_context.trace_state
 
         # The sampler decides whether to create a real or no-op span at the
         # time of span creation. No-op spans do not record events, and are not
@@ -747,7 +778,7 @@ class Tracer(trace_api.Tracer):
         # The sampler may also add attributes to the newly-created span, e.g.
         # to include information about the sampling result.
         sampling_result = self.source.sampler.should_sample(
-            parent_context, trace_id, name, attributes, links,
+            parent_span_context, trace_id, name, attributes, links,
         )
 
         trace_flags = (
@@ -755,9 +786,9 @@ class Tracer(trace_api.Tracer):
             if sampling_result.decision.is_sampled()
             else trace_api.TraceFlags(trace_api.TraceFlags.DEFAULT)
         )
-        context = trace_api.SpanContext(
+        span_context = trace_api.SpanContext(
             trace_id,
-            generate_span_id(),
+            self.source.ids_generator.generate_span_id(),
             is_remote=False,
             trace_flags=trace_flags,
             trace_state=trace_state,
@@ -766,10 +797,10 @@ class Tracer(trace_api.Tracer):
         # Only record if is_recording() is true
         if sampling_result.decision.is_recording():
             # pylint:disable=protected-access
-            span = Span(
+            span = _Span(
                 name=name,
-                context=context,
-                parent=parent_context,
+                context=span_context,
+                parent=parent_span_context,
                 sampler=self.source.sampler,
                 resource=self.source.resource,
                 attributes=sampling_result.attributes.copy(),
@@ -779,14 +810,17 @@ class Tracer(trace_api.Tracer):
                 instrumentation_info=self.instrumentation_info,
                 set_status_on_exception=set_status_on_exception,
             )
-            span.start(start_time=start_time)
+            span.start(start_time=start_time, parent_context=context)
         else:
-            span = trace_api.DefaultSpan(context=context)
+            span = trace_api.DefaultSpan(context=span_context)
         return span
 
     @contextmanager
     def use_span(
-        self, span: trace_api.Span, end_on_exit: bool = False
+        self,
+        span: trace_api.Span,
+        end_on_exit: bool = False,
+        record_exception: bool = True,
     ) -> Iterator[trace_api.Span]:
         try:
             token = context_api.attach(context_api.set_value(SPAN_KEY, span))
@@ -796,20 +830,24 @@ class Tracer(trace_api.Tracer):
                 context_api.detach(token)
 
         except Exception as error:  # pylint: disable=broad-except
-            if (
-                isinstance(span, Span)
-                and span.status is None
-                and span._set_status_on_exception  # pylint:disable=protected-access  # noqa
-            ):
-                span.set_status(
-                    Status(
-                        canonical_code=StatusCanonicalCode.UNKNOWN,
-                        description="{}: {}".format(
-                            type(error).__name__, error
-                        ),
-                    )
-                )
+            # pylint:disable=protected-access
+            if isinstance(span, Span):
+                if record_exception:
+                    span.record_exception(error)
 
+                if span.status is None and span._set_status_on_exception:
+                    span.set_status(
+                        Status(
+                            canonical_code=getattr(
+                                error,
+                                EXCEPTION_STATUS_FIELD,
+                                StatusCanonicalCode.UNKNOWN,
+                            ),
+                            description="{}: {}".format(
+                                type(error).__name__, error
+                            ),
+                        )
+                    )
             raise
 
         finally:
@@ -826,10 +864,15 @@ class TracerProvider(trace_api.TracerProvider):
         active_span_processor: Union[
             SynchronousMultiSpanProcessor, ConcurrentMultiSpanProcessor
         ] = None,
+        ids_generator: trace_api.IdsGenerator = None,
     ):
         self._active_span_processor = (
             active_span_processor or SynchronousMultiSpanProcessor()
         )
+        if ids_generator is None:
+            self.ids_generator = trace_api.RandomIdsGenerator()
+        else:
+            self.ids_generator = ids_generator
         self.resource = resource
         self.sampler = sampler
         self._atexit_handler = None