+++ exporter/opentelemetry-exporter-datadog/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Unreleased
 
+ - Make `SpanProcessor.on_start` accept parent Context
+  ([#1251](https://github.com/open-telemetry/opentelemetry-python/pull/1251))
+
+## Version 0.14b0
+
+Released 2020-10-13
+
 - Add support for span resource labels and service name 
 
 ## Version 0.12b0