+++ exporter/opentelemetry-exporter-otlp/CHANGELOG.md
@@ -2,8 +2,16 @@
 
 ## Unreleased
 
+gzip-compression-feature
 - Add Gzip compression for exporter
   ([#1141](https://github.com/open-telemetry/opentelemetry-python/pull/1141))
+
+## Version 0.14b0
+
+Released 2020-10-13
+
+- Add timestamps to OTLP exporter
+  ([#1199](https://github.com/open-telemetry/opentelemetry-python/pull/1199))
 - Update OpenTelemetry protos to v0.5.0
   ([#1143](https://github.com/open-telemetry/opentelemetry-python/pull/1143))